Fix WanSyncTrackingTest failure (#3741)
The test increments a sent event counter every time it sends a batch
to the target cluster and later asserts that the number of the sent
entries reported to MC equals to this value. This is done every time a
batch is sent. If a batch is sent multiple times, the entries in the
banch are counted multiple times. This is fixed in this commit by
counting the sent events only when the target cluster acknowledged the
batch. With this we count every batch only once even if for any reason
we need to resend the match.

Fixes #3286