Evaluate PMEM device check in runtime (#3750)
MemkindHeap cached the value of
"hazelcast.persistent.memory.check.disable" in a static field. If
`MemkindHeap` is loaded and initialized before a test attempts to set
the system property, the test will fail with the message `/tmp/.... is
not a persistent memory`. This is the case if `MemkindMallocTest` runs
in the same JVM before `MemkindPmemMallocTest` because both uses
`MemkindHeap`, but the former doesn't set the property.

This change solves this issue by evaluating the property during runtime
instead of reading the statically cached value. It shouldn't have any
performance overhead since the heaps are initialized at startup time
once and no new heap is created during runtime.

Fixes #3749