Fix MapWanSyncAPITest#testConcurrentSync (#3744)
Fix MapWanSyncAPITest#testConcurrentSync

The test was quite resource intensive because it was firing many sync
requests and replacing the map contents in parallel. It caused stalls
and other problems in the PR builder.

In reality, the test actually simply needs entries to be removed after
sync was started but before it completes. We achieve this by
synchronization instead of brute force by reusing a custom batch
publisher from another test.

Also, removed setting the batch sender via system property. It's quite
dangerous as it's JVM-wide and was used in parallel tests which
fortunately didn't fail. Plus, it the JVM argument wasn't even
cleaned up. Instead, we allow overriding the batch sender by overriding
a method on the batch publisher.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/3739