Fix ClusterHotRestartBackupTest.testClusterHotRestartBackup (#3771)
The member list is persisted to disk via an asynchronous process by
submitting a task to a specific queue. Hot restart backup is concurrent
to this process and will simply copy the member metadata as it is
observed at the time when the backup was triggered. There is then a
window in which a member has a more recent member list in-memory than it
has on-disk. If hot restart backup is triggered during this time, the
backup will contain a stale member list. The queue which contains the
task for persisting the member list is also processed during member
shutdown.

We then fix the test by first restarting the cluster, ensuring that the
latest member lists have been persisted, and calling hot restart backup
after the cluster has restarted.

Alternatively, we may introduce coordination by which the hot backup
process also forces processing of all tasks in the metadata persistence
queue. I don't think this is a right approach because the member list
persistence task might not be enqueued yet so flushing the queue might
not achieve anything, really. We'd have to coordinate somehow with both
the in-memory member list update and the task which persists it on disk.
But, again, this might not be good enough as the member list update is
asynchronous across the cluster, a member might have a stale member list
for some time. So adding coordination between the member list update
process and the hot backup process doesn't really achieve much but this
is open to discussion.

Also cleaned up some warnings.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/3576