Fix WAN sync statistics reporting (#3776)
The removeReplicationEvent method may be called concurrently for
different WAN sync operations. If one of these operations completes the
sync for a specific map partition, we may declare that the map partition
(or even the whole map) was synced before we actually count all of the
synced records. This is because the record counting is serialized in the
updateSerializingExecutor, while the remaining event count is counted
outside of this executor.

We now serialize all WAN statistics updates inside the executor so that
once we declare the map as synced, all of the other stats should be
correct.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/3302