Avoid MBean registrations for some cache tests (#3780)
Hot restart and split-brain ICache tests registered MBeans but didn't
unregister them. One approach to fix this is to properly destroy the
caches in an "After" method but this complicates tests as they must now
share the reference to the cache in a class field and sometimes the
cache is created multiple times in a single test and we cannot simply
destroy the cache. An example of such a test is a hot restart test which
asserts that a cache is present after restart.

Instead, we disable statistics which prevents registering the MBeans in
the first place. None of the affected tests need statistics.

Additionally, added assertions that the MBeans are cleaned up so in case
someone enables statistics, the test should start failing.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/3774