Perform zero size check in UnsafeMalloc
We rely on a specific behavior of our LibMalloc implementations, that
is, if the size to allocate is zero, we expect a NULL_ADDRESS returned.
But Unsafe is not a safe API, and it may just forward the call to the
underlying malloc()/realloc() implementation that behaves implementation
dependent in this case and may allocate minimum size chunks.

Then we fix this as checking the size in UnsafeMalloc and return
NULL_ADDRESS in the given cases. Reallocate also frees the chunk.

Fixes #3805