Force all cluster starts to wait for safe state in the WAN tests (#3818)
There are two variants of the start method in the WAN test framework.
The simple `startCluster` starts all not started members in a cluster
and immediately gives the control back the test once the members are
setup. The other variant is `startAndWaitForSafeState`, which waits for
all members to reach `SAFE` state and get their partitions warmed up.
For the WAN tests, there is no advantage of not doing the latter in
every case, while there is a disadvantage with doing the former. When a
member is about to join a cluster, it connects to every member it
can reach. In WAN tests this means the target cluster makes an attempt
to connect to the source cluster, which fails as their cluster-name is
different. If during this period when this connection is open WAN tries
to perform WAN consistency check, the check fails as the existing
connection is getting closed in a very short time. The consistency check
is performed outside of the WAN publishers and there is no retry logic
implemented. If the above described situation happens, the test fails
with this hard to debug reason. In the 4.1 we prevent this with making
all cluster starts to wait for safe clusters, hence to perform
anti-entropy operations on clusters in this state. In 4.2 we ideally
make the Merkle consistency check operation to go through the WAN
publishers that has the proper retry logic implemented.

Fixed #3180