Introduce mode for persistent-memory configuration (#3814)
Introduce config attribute for the operational mode of the persistent
memory. Two modes supported: `MOUNTED (FS DAX)` and `SYSTEM_MEMORY (KMEM
DAX)`. With this setting, the user can define to use PMEM in the KMEM
DAX
mode. Additionally, an `enabled` toggle is added to `persistent-memory`,
which makes it explicit when persisten memory is used. For legacy
reasons, if the `persistent-memory-directory` that was introduced in 4.0
is set, `enabled` is set to `true`, so the users already using
persistent memory don't need to change their configs. Otherwise, the
persistent memory should be explicitly enabled by setting `enabled` to
`true`, since the default is `false`.

The change also removes the hidden "hazelcast.hd.memkind.dax.kmem" system
argument and enable using KMEM DAX kind based on config instead.