Report initial WAN state in WAN stats (#3862)
WAN reports hardwired `STOPPED` publisher state in its statistics for
the existing publishers before the publishers replicated an event. This
is confusing since the initial publisher state is configurable. This
change addresses this issue by reporting the configured initial
publisher state for the existing but not yet initialized WAN publishers.

Fixes #3834