Add `boolean terminate` parameter to NodeExtension#beforeShutdown to.
This is necessary to gracefully shutdown the Jet jobs if the instance
shutdown gracefully.

Fix tests which add a MapConfig after starting the instance. Before the
merge, JetInstance#getConfig returns the JetConfig object which then
can be used to obtain the Config object. Normally this should be a
static object and shouldn't get mutated after the instance started. But
tests add MapConfig to it after the instance started.

After the merge, we add the MapConfig via HazelcastInstance#getConfig,
which returns DynamicConfigurationAwareConfig. Since we search for the
config using the static config object first, we find the pre-configured
MapConfig object named `__jet.*` instead of added MapConfig which is
named `__jet.snapshot.*`.

To fix this issue tests cast the Config object obtained via
HazelcastInstance#getConfig to `DynamicConfigurationAwareConfig` and
add the desired MapConfig to the
DynamicConfigurationAwareConfig#getStaticConfig.