Migrate in-flight WAN events (#3941)
* Migrate in-flight WAN events

In order to not lose WAN events, the WAN partition queues are getting
migrated when rebalancing the cluster. Despite this queue migration,
event loss was still possible, since until this commit, the WAN events
being replicated (added to an in-flight batch) were removed from these
queues, therefore, were not migrated. In some cases, this could lead
to losing the WAN events added to these batches. This change fixes this
problem by introducing a new WAN queue specilized for WAN. The queue
adds a third operation to the usual enqueue/dequeue operations: the
finalization operation. A dequeued WAN event remains in the queue until
that given WAN event and all other WAN events preceding it in the queue
are finalized. The finalization operation is called from the WAN callback
handlers ensuring that a WAN event is not removed from the queue
containing it until the target cluster ACKs the batch the event was
added to. This makes even these in-flight WAN operations migrated. To
prevent losing monotonicity, the already dequeued but not yet finalized
WAN events of the same keys are coalesced during migration. If the
target already received the event, there will be no state change. The
side-effect of this coalescing is that events of different keys on the
same partition may get reordered. Let's say key A and B on the same
partition receives updates in A1,B1,B2,A2 order. With coalescing their
replication order will be B2,A2. If the target cluster already received
the events, this change is irrelevant. If the events were not seen
previously by the target and it relies on their order, it can be a
problem. If the target already received the events in question, without
coalescing the A1,B1,B2,A2,A1,B1,B2,A2 update sequence would be
possible. In this change, coalescing is done unconditionally as
potentially changing the replication order seems to be a smaller issue
than temporarily resurrecting an already updated state. Optionally, this
coalescing can be configurable.

Since the WAN sync events are not going through the per partition WAN
queues, distinguishing the sync events and the replication events is
necessary. In order to do this, the InternalWanEvents are wrapped into
WrappedWanEvents, subclassed with a FinalizableWrappedWanEvent. This
adds a slight increase to the allocation cost of the replication.

The chosen queue implementation is a lock-based queue offering the same
consistency model that of the replaced LinkedBlockingQueue. There was an
alternative lock-free, on-the-typical-path scalable queue implementation
featuring collaborative concurrency, but that implementation got discarded.
In the mostly SPSC usage pattern of WAN (with the exception of clearing
the queue) the lock-based approach has shown somewhat better performance
results and also it's much easier to reason about its correctness in the
review. Since the queue is lock-based, clearing the queue from MC can block
the partition thread. This is a trade-off for the better performance with a
single producer and for simplicity. Also, this is not a change,
LinkedBlockingQueue also uses locks internally, clearing the queue requires
a full lock (locking on both head and tail). Later, if we change
the threading model of WAN, we may rethink using the other queue.

Fixes #3867