Implementation of HD Metadata Store (#3943)
Use HD Metadata Store to store all metadata off-heap as well. The
newly introduced Metadata Store is based on BinaryElasticHashMap.
We introduced a new type of record called HDMetadataRecord that
references the key and value part of the metadata.

Individual updates of the key/value parts of metadata are supported.
If both the key and value are NULL, the record is automatically removed.